feat: put adaptive sleep behind UCCL_RDMA_ADAPTIVE_SLEEP flag
